requirements.txt
python-telegram-bot
py-money
requests
pyYAML